Updated a couple projects to build and run in QS
Updated the Clipboard, Firefox, and WebSearch Modules to compile and run inside QS. Unfortunately, the Firefox source code is incomplete, and does not work like the module available from Blacktree.com.  The WebSearch Module also appears to be of an older version, so that may be a problem too.

It looks like (from looking at the info.plist file) that a new Firefox module could be modeled after the Safari Module, which is complete.

Also, just because a module compiles does not mean that QS will be able to run it. If the module compiles but does not load into QS or you can't set breakpoints in the module for debugging, then check the following items (there may be more than listed, but this is a good start)

1) Make sure the "Project Settings" are "Based on ..." release/debug, whichever is appropriate.
  1a) Select all project setting items and press the delete key to reset all items to the defaults (none of the project items should be bold).
2) Make sure the "Target Settings" are "Based on ..." QSPlugIn_Release/QSPlugIn_Debug, whichever is appropriate.
  2a) Make sure all the project items are at their default values (not bold) except for the follow items...
    2a1) [OTHER_LDFLAGS] "Other Linker Flags", set to "-bundle"
    2a2) [PRODUCT_NAME] "Product Name", set to "THE NAME OF YOUR PRODUCT" (ie. the output bindary file name, without the extension).
    2a3) Not needed, but if a user defined variable at the bottom called "ZERO_LINK" is set, then delete it.
3) Edit the "info.plist" file in the "Resources" group.
  3a) Find the key "Executable file" and make sure it's string is "$(EXECUTABLE_NAME)", without the quotes.
4) Under the "Frameworks" group make the following changes...
  4a) Delete "QSCore.framework", "QSEffects.framework", "QSFoundation.framework", "QSInterface.framework".  You're doing this because they probably are pointing of in la-la land.
  4b) From "Finder" open the directory "/Applications/Quicksilver.app/Contents/Frameworks".
  4c) Select all the frameworks (4 at this time) and drag to the Xcode project's Frameworks tree (under Groups and Files).
  4d) Select the newly added framework items in the tree and right click the mouse...Select the "Get info" context menu...and in the new dialog select "Relative to QSFramework" (you did set that setting in Xcode->preferences->Source Trees, didn't you).

With those settings out of the way, compile the plugin in debug mode, and if need be compile QS to make sure the plugin is added properly (check QS_APP_CONTENTS_FOLDER/plugins/  mine is /tmp/QS/build/Debug/Quicksilver.app/Contents/plugins/).  If it's there then start debugging QS as you normally would.

In a second Xcode session have your plugin project opened and set some breakpoints.  If they stick with a blue color then the plugin is probably loaded correctly in QS, with debug info.  If they turn yellow after setting, they recheck the items above (that list is not complete).  A good place to set initial breakpoints to see if everything is hooked up correctly is the init and initialize class methods.

Note - if you have multiple Xcode projects open then the breakpoint settings will be shared across the sessions.